requirements.txt
Django==2.2.6
djangorestframework==3.12.4
djoser==2.1.0

chardet~=4.0.0
